+++ modules/farm_rothamsted_experiment_research/src/Form/ProposalEntityForm.php
@@ -102,6 +102,7 @@ class ProposalEntityForm extends ResearchEntityForm {
           'design',
           'plan',
           'reviewer',
+          'study_id',
           'status',
           'status_notes',
         ],